modif multer routes put & delete

+++ Backend/controllers/book.js
@@ -1,4 +1,5 @@
 const Book = require("../models/Book.js");
+const fs = require("fs");
 
 exports.createBook = (req, res, next) => {
   const bookObject = JSON.parse(req.body.book);
@@ -15,7 +16,7 @@ exports.createBook = (req, res, next) => {
   book
     .save()
     .then(() => {
-      res.status(201).json({ message: "Objet enregistré !" });
+      res.status(201).json({ message: "Livre enregistré !" });
     })
     .catch((error) => {
       res.status(400).json({ error });
@@ -23,15 +24,59 @@ exports.createBook = (req, res, next) => {
 };
 
 exports.modifyBook = (req, res, next) => {
-  Book.updateOne({ _id: req.params.id }, { ...req.body, _id: req.params.id })
-    .then(() => res.status(200).json({ message: "Livre modifié !" }))
-    .catch((error) => res.status(400).json({ error }));
+  const bookObject = req.file
+    ? {
+        ...JSON.parse(req.body.book),
+        imageUrl: `${req.protocol}://${req.get("host")}/images/${
+          req.file.filename
+        }`,
+      }
+    : { ...req.body };
+
+  delete bookObject._userId;
+  Book.findOne({ _id: req.params.id })
+    .then((book) => {
+      if (book.userId != req.auth.userId) {
+        res.status(401).json({ message: "Non-autorisé" });
+      } else {
+        Book.updateOne(
+          { _id: req.params.id },
+          { ...bookObject, _id: req.params.id }
+        )
+          .then(() => {
+            res.status(200).json({ message: "Livre modifié" });
+          })
+          .catch((error) => {
+            res.status(401).json({ error });
+          });
+      }
+    })
+    .catch((error) => {
+      res.status(400).json({ error });
+    });
 };
 
 exports.deleteBook = (req, res, next) => {
-  Book.deleteOne({ _id: req.params.id })
-    .then(() => res.status(200).json({ message: "Livre supprimé !" }))
-    .catch((error) => res.status(400).json({ error }));
+  Book.findOne({ _id: req.params.id })
+    .then((book) => {
+      if (book.userId != req.auth.userId) {
+        res.status(401).json({ message: "Non autorisé" });
+      } else {
+        const filename = book.imageUrl.split("/images/")[1];
+        fs.unlink(`images/${filename}`, () => {
+          Book.deleteOne({ _id: req.params.id })
+            .then(() => {
+              res.status(200).json({ message: "Livre supprimé !" });
+            })
+            .catch((error) => {
+              res.status(401).json({ error });
+            });
+        });
+      }
+    })
+    .catch((error) => {
+      res.status(500).json({ error });
+    });
 };
 
 exports.getOneBook = (req, res, next) => {
@@ -46,19 +91,13 @@ exports.getAllBooks = (req, res, next) => {
     .catch((error) => res.status(400).json({ error }));
 };
 
-// exports.getBestRatings = (req, res, next) => {
-//   Book.find()
-//     .then((books) => {
-//       const bestBooks = books
-//         .sort((a, b) => b.averageRating - a.averageRating)
-//         .slice(0, 2);
-//       res.status(200).json(bestBooks);
-//     })
-//     .catch((error) => res.status(500).json({ error }));
-// };
-
 exports.getBestRatings = (req, res, next) => {
   Book.find()
-    .then((books) => res.status(200).json(books))
-    .catch((error) => res.status(400).json({ error }));
+    .then((books) => {
+      const bestBooks = books
+        .sort((a, b) => b.averageRating - a.averageRating)
+        .slice(0, 3);
+      res.status(200).json(bestBooks);
+    })
+    .catch((error) => res.status(500).json({ error }));
 };
